nan_high_neg

+++ Telescope/src/headers/common.h
@@ -1,4 +1,4 @@
 #pragma once
 
-typedef float real;  //Clarifies most of the variables of program.
+typedef double real;  //Clarifies most of the variables of program.
                      //In some variables there is type casting.
